refactor(#2918): add new util functions to processes module
